Created resuable PaginationControl component
